// app/lib/pdf-templates/types.ts
export interface LetterOfCredit {
  // Config
  title: string;
  fromLabel: string;
  dateLabel: string;
  toLabel: string;
  subjectLabel: string;
  bodyText1: string; // The main big paragraph
  bodyText2: string; // The "drawn under" paragraph
  closingText: string;
  
  // Data
  from: string;
  date: string;
  to: string;
  subject: string;
  drawnOn: string;
  amount: number;
  currency: string;
  accountOf: string;
  drawnUnder: string;
  commission: string;
  expirationDate: string;
  senderName: string;
  organizationName: string;
}